+++ packages/scripts/src/projects/background.ts
@@ -16,7 +16,7 @@ import {
 import semver_gt from 'semver/functions/gt';
 import semver_valid from 'semver/functions/valid';
 import { CommonProject } from './common';
-import { CXProject, definedProjects, ICourseProject, IcveMoocProject, ZHSProject, ZJYProject } from '..';
+import { CXProject, definedProjects, ICourseProject, IcveMoocProject, YKTProject, ZHSProject, ZJYProject } from '..';
 import { RenderScript } from '../render';
 import { SearchInfosElement } from '../elements/search.infos';
 import { $render } from '../utils/render';
@@ -925,6 +925,7 @@ export const BackgroundProject = Project.create({
 				const currentStudyScript = [
 					[CXProject.scripts.studyDispatcher, CXProject.scripts.study],
 					CXProject.scripts.work,
+					CXProject.scripts.autoRead,
 					ZHSProject.scripts['gxk-study'],
 					ZHSProject.scripts['xnk-study'],
 					ZHSProject.scripts.hike,
@@ -942,7 +943,8 @@ export const BackgroundProject = Project.create({
 					[ZJYProject.scripts.dispatcher, ZJYProject.scripts.study],
 					ZJYProject.scripts.work,
 					IcveMoocProject.scripts.study,
-					IcveMoocProject.scripts.work
+					IcveMoocProject.scripts.work,
+					YKTProject.scripts.ai
 				]
 					.map((m) => {
 						const url = window.location.href;